add unitest
